pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clawtail"
version = "0.1.0"
description = "ClawWithTail — Physical World AI Agent Platform"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "mcp>=1.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "apscheduler>=3.10.4",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "opencv-python>=4.9.0.80",
    "ultralytics>=8.2.0",
    "bleak>=0.21.1",
    "pyserial>=3.5",
    "python-slugify>=8.0.4",
    "tavily-python>=0.3.3",
    "jsonschema>=4.22.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-mock>=3.14.0",
    "coverage>=7.5.0",
]

[project.scripts]
clawtail = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = [
    "core",
    "tools",
    "adapters",
    "env",
    "cli",
    "tasks",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
